requirements.txt
opencv-python
packaging
pdf2image
pillow
pip
pytesseract
setuptools
wheel
Flask
SQLAlchemy
flask_sqlalchemy
pytest
python-dotenv
flask_cors
pymysql